+++ src/stories/pages/404Page/ErrorPage.stories.tsx
@@ -4,12 +4,10 @@ import { Meta, StoryObj } from '@storybook/react';
 import {
   reactRouterNestedAncestors,
   reactRouterParameters,
-  withRouter,
 } from 'storybook-addon-react-router-v6';
 
 const meta: Meta<typeof ErrorPage> = {
   component: ErrorPage,
-  decorators: [withRouter],
   parameters: {
     reactRouter: reactRouterParameters({
       routing: reactRouterNestedAncestors(<PageLayout />),
