+++ src/stories/pages/BoardEnterPage/BoardEnterPage.stories.tsx
@@ -4,12 +4,10 @@ import { Meta, StoryObj } from '@storybook/react';
 import {
   reactRouterNestedAncestors,
   reactRouterParameters,
-  withRouter,
 } from 'storybook-addon-react-router-v6';
 
 const meta: Meta<typeof BoardEnterPage> = {
   component: BoardEnterPage,
-  decorators: [withRouter],
   parameters: {
     reactRouter: reactRouterParameters({
       routing: reactRouterNestedAncestors(<PageLayout />),
