+++ src/stories/pages/BoardPage/BoardPage.stories.tsx
@@ -4,12 +4,10 @@ import { Meta, StoryObj } from '@storybook/react';
 import {
   reactRouterNestedAncestors,
   reactRouterParameters,
-  withRouter,
 } from 'storybook-addon-react-router-v6';
 
 const meta: Meta<typeof BoardPage> = {
   component: BoardPage,
-  decorators: [withRouter],
   parameters: {
     reactRouter: reactRouterParameters({
       routing: reactRouterNestedAncestors(<PageLayout />),
