+++ src/stories/pages/MyPage/MyPage.stories.ts
@@ -1,10 +1,8 @@
 import MyPage from '@/pages/MyPage';
 import { Meta, StoryObj } from '@storybook/react';
-import { withRouter } from 'storybook-addon-react-router-v6';
 
 const meta: Meta<typeof MyPage> = {
   component: MyPage,
-  decorators: [withRouter],
   argTypes: {},
 };
